\name{bitcoin_price}
\alias{bitcoin_price}
\docType{data}
\title{
Bitcoin price between January 1, 2018 and February 4, 2025.
}
\description{
A time series of size 2,591.
}
\usage{data("bitcoin_price")}
\format{
  \describe{
    \item{\code{Open.time}}{Day.}
    \item{\code{Open}}{Opening price.}
    \item{\code{High}}{Highest price.}
    \item{\code{Low}}{Lowest price.}
    \item{\code{Close}}{Closing price.}
  }
}
\source{
Binance.
}
\examples{
data(bitcoin)
}
\keyword{datasets}
