\name{boeing_price}
\alias{boeing_price}
\docType{data}
\title{
Boeing stock prices between March 20, 2015 and March 19, 2025.
}
\description{
A time series of size 2,515.
}
\usage{data("boeing_price")}
\format{
  \describe{
    \item{\code{Date}}{Day.}
    \item{\code{Open}}{Opening price.}
    \item{\code{High}}{Highest price.}
    \item{\code{Low}}{Lowest price.}
    \item{\code{Close}}{Closing price.}
  }
}
\source{
MarketWatch.
}
\examples{
data(boeing_price)
}
\keyword{datasets}
