\name{french_motor}
\alias{french_motor}
\docType{data}
\title{
French Motor TPL Insurance Claims Data dataset
}
\description{
A data set of 24,944 insurance policyholders with strictly positives claim amounts.
}
\usage{data("french_motor")}
\format{
  \describe{
    \item{\code{IDpol}}{ID of the policyholder.}
    \item{\code{ClaimNb}}{Number of claims.}
    \item{\code{Exposure}}{Exposure.}
    \item{\code{VehPower}}{Vehicule power.}
    \item{\code{VehAge}}{Vehicule age.}
    \item{\code{DrivAge}}{Driver age.}
    \item{\code{BonusMalus}}{Bonus malus.}
    \item{\code{VehBrand}}{Vehicule brand.}
    \item{\code{VehGas}}{Vehicule gas.}
    \item{\code{Area}}{Area.}
    \item{\code{Density}}{Density.}
    \item{\code{Region}}{Region.}
    \item{\code{Claim_Amount}}{Total claim amount.}
  }
}
\source{
Unknown private insurer.
}
\examples{
data(french_motor)
}
\keyword{datasets}
